should be bms, not aux bms
