--- EpsilonCCS/Src/CanParsers/AuxBmsCanParser.c
@@ -1,31 +1,0 @@
-#include "AuxBmsCanParser.h"
-
-void parseBmuCanMessage(uint32_t stdId, uint8_t data[8])
-{
-
-}
-
-void parseBmuHeartbeat(uint8_t data[8])
-{
-
-}
-
-void parseStartupInfo(uint8_t data[8])
-{
-
-}
-
-void parsePackInfo(uint8_t data[8])
-{
-
-}
-
-void parseErrorsTempInfo(uint8_t data[8])
-{
-
-}
-
-void parseCellVoltages(uint8_t data[8])
-{
-
-}
