+++ EpsilonCCS/Src/CanParsers/CanParser.c
@@ -28,5 +28,8 @@
 
 void parseCanMessage(uint32_t stdId, uint8_t data[8])
 {
-
+	if(stdId & BMS_CAN_MASK == BMS_CAN_ID)
+	{
+		parseBmuCanMessage(stdId, data);
+	}
 }
